sieve: fix respectinbox=n
When ${global.respectinbox} is "y", the inbox flag is preserved and
starred messages are _also_ moved to the inbox if the user is in To
or Cc.

When ${global.respectinbox} is "n", the inbox flag is applied to
all non-starred messages.  Starred messages are only moved to the
inbox if the user is in To or Cc.